+++ bikram/__init__.py
@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 from .bikram import *   # NOQA
 
 __author__ = """keshaB Paudel"""
 __email__ = 'self@keshab.net'
-__version__ = '0.1.3'
+__version__ = '1.0.0'