+++ docs/history.rst
@@ -1,1 +1,23 @@
-.. include:: ../HISTORY.rst
+=======
+History
+=======
+
+1.0.0 (2018-03-15)
+------------------
+
+* Added usage guide and module docs.
+* Added :code:`pipenv` support.
+* Removed :code:`tox`.
+* Removed support for python versions older than 3.6.
+
+
+0.1.2 (2016-11-13)
+------------------
+
+* Minor test fixes.
+
+
+0.1.0 (2016-11-13)
+------------------
+
+* First release on PyPI.